Add a SessionTicker component that automatically refreshes the token when the page is open.
