+++ src/App.tsx
@@ -1,16 +1,17 @@
 import { useEffect } from "react";
 import { BrowserRouter, Routes as RouterRoutes, Route } from "react-router-dom";
-import { Provider, useDispatch } from "react-redux";
+import { Provider } from "react-redux";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 import GitInfo from "generated/gitInfo";
 
-import { AppDispatch, store } from "store";
+import { store } from "store";
 import { Routes } from "constants/settings";
 import { PrivateRoute } from "components/PrivateRoute";
 import { InnerPage } from "components/InnerPage";
 import { UserSession } from "components/UserSession";
 import { GlobalBanner } from "components/GlobalBanner";
+import { SessionTicker } from "components/SessionTicker";
 
 import { SignIn } from "pages/SignIn";
 import { MFAuth } from "pages/MFAuth";
@@ -39,7 +40,6 @@ import { Unauthorized } from "pages/Unauthorized";
 import { SigninOidc } from "pages/Redirect";
 
 import "styles.scss";
-import { refreshSessionToken } from "helpers/refreshSessionToken";
 
 const queryClient = new QueryClient({
   defaultOptions: {
@@ -51,17 +51,6 @@ const queryClient = new QueryClient({
 });
 
 export const App = () => {
-  const dispatch: AppDispatch = useDispatch();
-  useEffect(() => {
-    // Function to call refreshSessionToken every minute
-    const ticker = setInterval(() => {
-      refreshSessionToken(dispatch);
-    }, 10000); // 60,000ms = 1 minute
-
-    // Cleanup function to clear the interval when component unmounts
-    return () => clearInterval(ticker);
-  }, [dispatch]);
-
   useEffect(() => {
     // Git commit hash
     console.log("current commit hash: ", GitInfo.commitHash);
@@ -70,6 +59,7 @@ export const App = () => {
 
   return (
     <Provider store={store}>
+      <SessionTicker />
       <QueryClientProvider client={queryClient}>
         <UserSession />
         <BrowserRouter>
